extract delete task dialog
