+++ site/src/modules/tasks/TasksSidebar/TasksSidebar.stories.tsx
@@ -1,9 +1,9 @@
 import { MockTasks, MockUserOwner, mockApiError } from "testHelpers/entities";
-import { withAuthProvider, withGlobalSnackbar } from "testHelpers/storybook";
+import { withAuthProvider } from "testHelpers/storybook";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { API } from "api/api";
 import { MockUsers } from "pages/UsersPage/storybookData/users";
-import { expect, spyOn, userEvent, waitFor, within } from "storybook/test";
+import { spyOn, userEvent, within } from "storybook/test";
 import { reactRouterParameters } from "storybook-addon-remix-react-router";
 import { TasksSidebar } from "./TasksSidebar";
 
@@ -93,7 +93,7 @@ export const OpenOptionsMenu: Story = {
 	},
 };
 
-export const DeleteTaskDialog: Story = {
+export const OpenDeleteDialog: Story = {
 	beforeEach: () => {
 		spyOn(API.experimental, "getTasks").mockResolvedValue(MockTasks);
 	},
@@ -111,52 +111,6 @@ export const DeleteTaskDialog: Story = {
 				name: /delete/i,
 			});
 			await userEvent.click(deleteButton);
-		});
-	},
-};
-
-export const DeleteTaskSuccess: Story = {
-	decorators: [withGlobalSnackbar],
-	parameters: {
-		chromatic: {
-			disableSnapshot: false,
-		},
-	},
-	beforeEach: () => {
-		spyOn(API.experimental, "getTasks").mockResolvedValue(MockTasks);
-		spyOn(API.experimental, "deleteTask").mockResolvedValue();
-	},
-	play: async ({ canvasElement, step }) => {
-		const body = within(canvasElement.ownerDocument.body);
-		const canvas = within(canvasElement);
-
-		await step("Open menu", async () => {
-			const optionButtons = await canvas.findAllByRole("button", {
-				name: /task options/i,
-			});
-			await userEvent.click(optionButtons[0]);
-		});
-
-		await step("Open delete dialog", async () => {
-			const deleteButton = await body.findByRole("menuitem", {
-				name: /delete/i,
-			});
-			await userEvent.click(deleteButton);
-		});
-
-		await step("Confirm delete", async () => {
-			const confirmButton = await body.findByRole("button", {
-				name: /delete/i,
-			});
-			await userEvent.click(confirmButton);
-			await step("Confirm delete", async () => {
-				await waitFor(() => {
-					expect(API.experimental.deleteTask).toHaveBeenCalledWith(
-						MockTasks[0].workspace.owner_name,
-						MockTasks[0].workspace.id,
-					);
-				});
-			});
 		});
 	},
 };
